Source should be C, not C++
Old habit, made C++ files. This commit is mostly
renaming the previous versions.

+++ oblig2/src/binaryformat.c
@@ -7,7 +7,7 @@ void read_matrix_binaryformat (char* filename, double*** matrix, int* num_rows, 
     fread (num_cols, sizeof(int), 1, fp);
 
     alloc_matrix(matrix, (*num_rows), (*num_cols));
-    for (size_t i = 0; i < (*num_rows); i++) {
+    for (int i = 0; i < (*num_rows); i++) {
         fread((*matrix)[i], sizeof(double), (*num_cols), fp);
     }
     fclose(fp);
@@ -18,7 +18,7 @@ void write_matrix_binaryformat (char* filename, double** matrix, int num_rows, i
     FILE *fp = fopen (filename,"wb");
     fwrite (&num_rows, sizeof(int), 1, fp);
     fwrite (&num_cols, sizeof(int), 1, fp);
-    for (size_t i = 0; i < num_rows; i++) {
+    for (int i = 0; i < num_rows; i++) {
         fwrite (matrix[i], sizeof(double), num_cols, fp);
     }
     fclose (fp);
